First working version
main.py contains a simple run script that takes in a folder containing programs
(benchmarks) to run and a folder to store log files in. It launches the
programs sequentially and records the gathered data into a log file after the
program completes.

program_profile.py contains the profiling framework. Before each program is
launched, a separate process is started that records the measured power every
quantum (defaults to 0.5 seconds). A timestamp is taken when the program
begins; a timestamp is taken when the program ends. The separate process is
signaled to end, and it sends back all recorded power measurements (each of
which is also timestamped).

To extend to new machines, a new class that inherits ProgramProfile should be
created that implements the read_power function.