Added ability to specify subset of programs
Added an optional command line argument "config" that allows the
user to specify a file that contains a subset of programs in a
given folder to run.

The format of the file should be one program per line -- e.g. for
programs 'foo' and 'bar' in a folder containing 'foo', 'bar', and
'awk', the config file would look like this:

    foo
    bar

Note that as of this commit this functionality is NOT tested.