Added ability to set thread counts to run
Added optional command line argument to select the number of
threads to run (defaults to 1). For example,
  --thread_counts=1,2,4
will run all programs for 1, 2, and 4 threads (if supported). It
does this by setting the OMP_NUM_THREADS environment variable.

Also added error messages for the Xeon and X-Gene profiles; if the
power readings are all 0 (indicating that the power driver is not
installed), then the profiling framework fails early (as a precaution).

This is tested and working for the Xeon profile only. Additionally,
the --config option is now tested for the Xeon profile.